+++ server/routes/index.js
@@ -8,12 +8,16 @@ const items       = require('./items.js');
 const categories  = require('./categories.js');
 const conditions  = require('./conditions.js');
 const itemStatus  = require('./itemStatus.js');
+const login       = require('./login.js');
+const logout      = require('./logout.js');
 
 route.use('/register', register);
+route.use('/login', login);
 route.use('/users', users);
 route.use('/items', items);
 route.use('/categories', categories);
 route.use('/conditions', conditions);
 route.use('/itemStatus', itemStatus);
+route.use('/logout', logout);
 
 module.exports = route;
