+++ server/routes/logout.js
@@ -1,0 +1,16 @@
+const express                 = require('express');
+const app                     = express();
+const passport                = require('passport');
+const bcrypt                  = require('bcrypt');
+const LocalStrategy           = require('passport-local').Strategy;
+const saltRounds              = 12;
+const route                   = express.Router();
+const db                      = require('../models');
+
+route.get('/', (req,res) =>{
+  req.logout();
+  res.sendStatus(200);
+});
+
+
+module.exports = route;
